% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{assessment_parser}
\alias{assessment_parser}
\title{Parses Assessment String}
\usage{
assessment_parser(data, labels = NA, verbose = F)
}
\arguments{
\item{data}{assessment string}

\item{labels}{labels string (starts from OpenMind version 2.911)}

\item{verbose}{enable verbose printing}
}
\description{
Debugging Examples
}
\details{
data <- om_user_data %>%
filter(AssessmentVersion >= 5) %>%
# filter(PostAssessmentComplete == 1) %>%
pull(AssessmentAnswers) %>% .[[2000]]

error_dat %>%
  filter(errors != "No Error") -> ss

error_dat %>%
  filter(errors != "No Error")  %>%
  pull(OMID) %>% .[[3]]

data <- error_dat %>%
filter(errors != "No Error")  %>%
pull(AssessmentAnswers) %>% .[[3]]


data <- om_user_data %>%
filter(OMID %in% error_ids)  %>%
  mutate(OMID = factor(OMID, levels = error_ids)) %>%
  arrange(OMID) %>%
pull(AssessmentAnswers) %>%
  .[1]
#
assessment_parser(data)

 base_dat$FeedbackAnswers %>%
  .[[13]]
}
