% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{bind_questions}
\alias{bind_questions}
\title{Conducts t-tests and calculates Cohen's d for Q14 and Q17 seperately}
\usage{
bind_questions(.data, waves)
}
\arguments{
\item{...}{other arguments passed to summarize_comparison}
}
\description{
This is lower-level function that belongs to om_summarize_comparisons. This function is not meant to be used outside of om_summarize_comparisons.
}
