% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{calc_correct}
\alias{calc_correct}
\title{This calculates the correct score for each step}
\usage{
calc_correct(StepsComplete, StepsScores, StepQuestionTotals)
}
\arguments{
\item{StepQuestionTotals}{How many Question were there in total?}

\item{StepScomplete}{(0/1) Was Step completed?}

\item{Stepscores}{What was the score on the step?}
}
\description{
This calculates the correct score for each step
}
