% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{calc_ih}
\alias{calc_ih}
\title{Creates Intellectual Humility variable}
\usage{
calc_ih(final_dat, wave)
}
\arguments{
\item{final_dat}{Assessment data from AirTable}

\item{wave}{Specify wave ("Pre", "Post" or "FollowUp")}
}
\description{
This is lower-level function that belongs to om_construct measure. This function is not meant to be used outside of om_construct_measure.
Creates the following measure
\itemize{
  \item Q18: Intellectual Humility
}
Function automatically accounts for Assessment Version 4 and 5/5.1.
}
