% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{calc_measures}
\alias{calc_measures}
\title{Creates Target variables}
\usage{
calc_measures(final_dat, wave)
}
\arguments{
\item{final_dat}{Assessment data from AirTable}

\item{wave}{Specify wave ("Pre", "Post" or "FollowUp")}
}
\description{
This is lower-level function that belongs to om_construct measure. This function is not meant to be used outside of om_construct_measure.
Creates the following measures
\itemize{
  \item Q18: Intellectual Humility
  \item Q19: Perspective-Taking
}
Function automatically accounts for Assessment Version 4 and 5/5.1.
}
