% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_helpers.R
\name{chr_to_date}
\alias{chr_to_date}
\title{This function removes characters in AirTable and turns them to date}
\usage{
chr_to_date(date)
}
\arguments{
\item{date}{date}
}
\description{
This function removes characters in AirTable and turns them to date
}
