% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{chr_to_datetime}
\alias{chr_to_datetime}
\title{Parse a Character and turn it to Datetime}
\usage{
chr_to_datetime(date)
}
\arguments{
\item{date}{date (chr)}
}
\description{
Parse a Character and turn it to Datetime
}
