% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{coalesce_join}
\alias{coalesce_join}
\title{Coalescing joints}
\usage{
coalesce_join(x, y, by = NULL, suffix = c(".x", ".y"),
  join = dplyr::full_join, ...)
}
\arguments{
\item{x}{left hand side data}

\item{y}{right hand side data}

\item{by}{key variable}

\item{suffix}{suffix of data (default: .x and .y)}

\item{join}{dplyr join function}

\item{...}{additional argument to join}
}
\description{
Source: https://alistaire.rbind.io/blog/coalescing-joins/
}
