% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{do_if}
\alias{do_if}
\title{A pipable if statement}
\usage{
do_if(.data, condition, call)
}
\arguments{
\item{.data}{tibble}

\item{condition}{logical test}

\item{call}{a formula descibing a pipe to be evaluated if condition is \code{code}}
}
\description{
This function allows to create an if statement that can be used within a pipable workflow
}
\examples{
any_condition <- TRUE

mtcars \%>\%
do_if(any_condition, ~{
.x \%>\%
 dplyr::filter(cyl == 6) \%>\%
 dplyr::mutate(x = disp > 170)
})
}
