% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_helpers.R
\name{get_colnames}
\alias{get_colnames}
\title{This function assigns colnames}
\usage{
get_colnames(col_length, b2q, b1q, b3q)
}
\arguments{
\item{col_length}{how long is the column}

\item{b2q}{logical, was was B2 found?}

\item{b1q}{logical, was was B1 found?}

\item{b3q}{logical, was was B3 found?}
}
\description{
This function is necessary because variable names have to be guessed
Becomes obsolute in OpenMind version 2.911
}
