% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{get_error_diag}
\alias{get_error_diag}
\title{Diagnostic Dataset}
\usage{
get_error_diag(parsed_data)
}
\arguments{
\item{parsed_data}{Parsed dataset that is an output from `safely`}
}
\description{
Get diagnostic data from parser
}
