% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{if_flow}
\alias{if_flow}
\title{Pipable if workflow}
\usage{
if_flow(.x, if_true, pipe_work)
}
\arguments{
\item{.x}{dataset}

\item{if_true}{logical statement}

\item{pipe_work}{A function that will be performed on \code{TRUE}}
}
\description{
This function allows to create if statements within a pipe workflow
}
