% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{install_openmindR}
\alias{install_openmindR}
\title{Install newest openmindR version from GitHub}
\usage{
install_openmindR(...)
}
\arguments{
\item{...}{arguments to install_github}
}
\description{
This function installs the newest openmindR version from GitHub
}
