% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifehacks.R
\name{lh_reason}
\alias{lh_reason}
\title{Count LifeHacks Reason}
\usage{
lh_reason(lh_data, lifehack, lifehackcomplete, reason, step)
}
\arguments{
\item{lh_data}{a dataset that has gone through \code{\link{om_parse_lifehacks}}}

\item{lifehack}{Specify LifeHacks from which step you want to count}

\item{lifehackcomplete}{Specify LifeHacksComplete from which step you want to count}

\item{reason}{Specify LifeHacksReason from which step you want to count}

\item{step}{Swhich step are we talking about}
}
\description{
Happens for each Step.
}
