% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifehacks.R
\name{lh_useful}
\alias{lh_useful}
\title{Count LifeHacks Useful}
\usage{
lh_useful(lh_data, lifehack, lifehackcomplete, useful, step)
}
\arguments{
\item{lh_data}{a dataset that has gone through \code{\link{om_parse_lifehacks}}}

\item{lifehack}{Specify LifeHacks from which step you want to count}

\item{lifehackcomplete}{Specify LifeHacksComplete from which step you want to count}

\item{useful}{Specify LifeHacksUseful from which step you want to count}

\item{step}{Swhich step are we talking about}
}
\description{
Happens for each Step.
}
