% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{mailchimp_member_list}
\alias{mailchimp_member_list}
\title{Get MailChimp Member List}
\usage{
mailchimp_member_list(list_id = "c7e9a42a8f")
}
\arguments{
\item{list_id}{unique id that denotes MailChimp list}
}
\description{
This function wraps the python module "mailchimp3" to get all members of a list through the Mailchimp API
}
