% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{om_clean_par}
\alias{om_clean_par}
\title{Performs cleaning on ParticipantProgress data from AirTable}
\usage{
om_clean_par(dat.par, ...)
}
\arguments{
\item{dat.par}{(0/1) ParticipantProgress data from AirTable}

\item{...}{Arguments for select to get more from ParticipantProgress}
}
\description{
Creates the following measures
\itemize{
  \item StepTimes1 to StepTimes5: Duration in minutes to complete a step
  \item StepCorrect1 to StepCorrect5: Percentage of correct answers for each step
  \item FeedbackAnswers Q1 to Q5
}
}
