% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{om_compare}
\alias{om_compare}
\title{Conduct t-tests and calculate Cohen's d}
\usage{
om_compare(gathered_dat, compare = c("PrePost", "PreFollow",
  "PrePostFollow"))
}
\arguments{
\item{gathered_dat}{Assessment data as long format}

\item{compare}{With the `compare` argument you can specify either \code{"PrePost"}, \code{"PreFollow"} or both \code{c("PrePost", "PreFollow")} comparisons (the latter is the default).}
}
\description{
This is lower-level function that belongs to om_summarize_comparisons. This function is not meant to be used outside of om_summarize_comparisons.
}
