% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{om_count_}
\alias{om_count_}
\title{String count (because it will be deprecated from dplyr)}
\usage{
om_count_(x, vars, wt = NULL, sort = FALSE)
}
\arguments{
\item{x}{dataseet}

\item{vars}{which variable to count}

\item{wt}{weights}

\item{sort}{Boolean. Sort by count.}
}
\description{
This is just a helper function because it will be deprecated from dplyr at some point and we would like to keep it.
}
