% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{om_download_at}
\alias{om_download_at}
\title{Download AirTable data}
\usage{
om_download_at(key, tables = c("AccessCodes", "ParticipantProgress",
  "InstructorSurveyV2", "TechnicalInquiries"), clean = F, file = NULL,
  v6.1 = F)
}
\arguments{
\item{key}{key for AirTable API}

\item{tables}{specify which tables you want to download}
}
\value{
a list with (several) dataframe(s)
}
\description{
This function downloads current AirTable data
}
\section{Currently allowed input}{

\itemize{
\item \code{"AssessmentV6"}
\item \code{"AssessmentV7"}
\item \code{"AccessCodes"}
\item \code{"ParticipantProgress"}
\item \code{"ParticipantProgress2"}
\item \code{"InstructorSurveyV2"}
\item \code{"TechnicalInquiries"}
}
}

