% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{om_dummy_individ}
\alias{om_dummy_individ}
\title{Code a dummy variables for Individual users}
\usage{
om_dummy_individ(assessment)
}
\arguments{
\item{assessment}{assessment data}
}
\description{
This function creates the dummy variables from AccessCode
\itemize{
  \item ut_individ: Access Codes users (0) v. individual users (1)
}
}
\examples{
assessmentv7 \%>\%
  om_dummy_individ()
}
