% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{om_dummy_nonstraight}
\alias{om_dummy_nonstraight}
\title{Code a dummy variable for straight/non-straight}
\usage{
om_dummy_nonstraight(assessment)
}
\arguments{
\item{assessment}{assessment data}
}
\description{
This function creates a dummy variable from D5 (Sexuality) called \code{sex_nonstraight} and codes people who identify as heterosexual as 0 and everyone else as 1.
}
\examples{
assessmentv7 \%>\%
  om_dummy_nonstraight()
}
