% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{om_dummy_nonwhite}
\alias{om_dummy_nonwhite}
\title{Code a dummy variable for white/nonwhite}
\usage{
om_dummy_nonwhite(assessment)
}
\arguments{
\item{assessment}{assessment data}
}
\description{
This function creates a dummy variable from D3 (Race) called \code{race_nonwhite} and codes people who identify only as white as 0 and everyone else as 1.
}
\examples{
assessmentv7 \%>\%
  om_dummy_nonwhite()
}
