% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{om_dummy_ut}
\alias{om_dummy_ut}
\title{Code a dummy variables for UserType variables}
\usage{
om_dummy_ut(assessment)
}
\arguments{
\item{assessment}{assessment data}
}
\description{
This function creates the dummy variables from UserType
\itemize{
  \item ut_college_individ: College student (0) v. individual user (1)
  \item ut_corp_individ: Corp (0) v. individual user (1)
  \item ut_college_corp: College student (0) v. corp (1)
  \item ut_individ: Access Codes users (0) v. individual users (1)
}
}
\examples{
assessmentv7 \%>\%
  om_dummy_ut()
}
