% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{om_gather}
\alias{om_gather}
\title{Gather OpenMind data}
\usage{
om_gather(.data, which_strings = openmindR::q_c_strings)
}
\arguments{
\item{.data}{Assessment data}

\item{which_strings}{a string indicating which variables should be parsed out (default is \code{openmindR::q_c_strings})}
}
\description{
This function will turn Assessment data into long format and
creates the following variables:
\itemize{
  \item Question: Q1Pre, Q2Pre, Q3Pre etc.
  \item Time: Pre, Post, or FollowUp
  \item Response: Values of the Question
  \item variable_code: Q1, Q2, Q3 etc.
}
Function takes strings in the form of \code{"Q1P|Q2P"} for Pre- and Post Questions 1 & 2
}
\section{Workflow}{

\code{\link{om_filter_data}} -> \code{\link{om_rescale}} -> \code{\link{om_clean_ppol}} -> \code{\link{om_construct_measures}} -> \code{\link{remove_dups}} -> \code{\link{om_gather}}
}

\examples{
# Turn AffPol1 variable from v7 into long format
assessmentv7 \%>\%
  om_gather(which_strings = "AffPol1")

# Turn v7 into long format
assessmentv7 \%>\%
  om_gather(which_strings = v7_var_strings)
}
