% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{om_label_stats}
\alias{om_label_stats}
\title{Prepare paired data for plot with within subject error term}
\usage{
om_label_stats(gathered_dat, aversion = "All")
}
\arguments{
\item{gathered_dat}{Assessment data as long format}

\item{aversion}{AssessmentVersion should be one of \code{"V4"}, \code{"V5/V5.1"} or \code{"All"}}
}
\description{
This function calculates several measures for plotting
}
