% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifehacks.R
\name{om_lifehacks_summary2}
\alias{om_lifehacks_summary2}
\title{Parse LifeHacks}
\usage{
om_lifehacks_summary2(lh_data, ut_filter = "All")
}
\arguments{
\item{cleaned_dat}{a dataset which includes LifeHacks to be parsed and checked}
}
\description{
Parse LifeHacks and make sure that they all make sense by checking them.
}
