% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{om_lmer}
\alias{om_lmer}
\title{Run mixed models (with interactions)}
\usage{
om_lmer(mod, type = "int", mod_transform = "minmax")
}
\arguments{
\item{mod}{a fitted model}

\item{type}{what kind of model (currently only accepts \code{"int"} for interactions)}

\item{mod_transform}{Indicates which values of the moderator variable should be
used when plotting interaction terms (i.e. \code{type = "int"}). \describe{
\item{\code{"minmax"}}{(default) minimum and maximum values (lower and
upper bounds) of the moderator are used to plot the interaction between
independent variable and moderator(s).} \item{\code{"meansd"}}{uses the
mean value of the moderator as well as one standard deviation below and
above mean value to plot the effect of the moderator on the independent
variable (following the convention suggested by Cohen and Cohen and
popularized by Aiken and West (1991), i.e. using the mean, the value one
standard deviation above, and the value one standard deviation below the
mean as values of the moderator, see
\href{http://www.theanalysisfactor.com/3-tips-interpreting-moderation/}{Grace-Martin
K: 3 Tips to Make Interpreting Moderation Effects Easier}).}
\item{\code{"zeromax"}}{is similar to the \code{"minmax"} option, however,
\code{0} is always used as minimum value for the moderator. This may be
useful for predictors that don't have an empirical zero-value, but absence
of moderation should be simulated by using 0 as minimum.}
\item{\code{"quart"}}{calculates and uses the quartiles (lower, median and
upper) of the moderator value.} \item{\code{"all"}}{uses all values of the
moderator variable.} }}
}
\description{
This function performs a mixed model and gives back some neat info
}
