% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{om_mix_complete}
\alias{om_mix_complete}
\title{Run and plot mixed effects model}
\usage{
om_mix_complete(experiment, title)
}
\arguments{
\item{experiment}{Specify a dataset with experimental data}

\item{title}{specify a title and y-label for the plot}
}
\description{
This function allows to run and plot a mixed model. It makes use of both \code{\link{om_mix_models}} and \code{\link{om_mix_plot}} (currently only works on Ann Miller experimental data)
}
