% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{om_mix_plot}
\alias{om_mix_plot}
\title{Plot mixed effects model}
\usage{
om_mix_plot(effects_dat, tidy_dat = NULL, var_label, show_stats = T)
}
\arguments{
\item{effects_dat}{\code{effects_dat} is a dataset produced by \code{\link{om_mix_models}} and supplies the marginal effects of the model}

\item{tidy_dat}{\code{tidy_dat} is a dataset produced by \code{\link{om_mix_models}} and supplies the pseudo cohen's d for plotting}

\item{var_label}{supply a character that is plotted as title and y-axis}

\item{show_stats}{Show statistics on the bottom right. Only possible if you supply \code{tidy_dat}}
}
\description{
This function plots the results of mixed models (Currently only works on Ann Miller experimental data)
}
