% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifehacks.R
\name{om_parse_lifehacks}
\alias{om_parse_lifehacks}
\title{Parse LifeHacks}
\usage{
om_parse_lifehacks(cleaned_dat)
}
\arguments{
\item{cleaned_dat}{a dataset which includes LifeHacks to be parsed and checked}
}
\description{
Parse LifeHacks and make sure that they all make sense by checking them.
}
