% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{om_rescale}
\alias{om_rescale}
\title{Rescale Variables}
\usage{
om_rescale(.data)
}
\arguments{
\item{.data}{Assessment data}
}
\description{
This function rescales variables from 0 to 1. Q1 and Q2 is divided by 100 and Q3 - Q12 as well as C1 - C3 (divided by 6).
}
