% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{om_reverse_code}
\alias{om_reverse_code}
\title{Reverse Code items}
\usage{
om_reverse_code(assessment)
}
\arguments{
\item{assessment}{v7 data}
}
\description{
Reverse codes items and adds them at the end of the dataset with a "_Rev" at the end.
}
\examples{
assessmentv7 \%>\%
  om_reverse_code()
}
