% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{om_strsplit}
\alias{om_strsplit}
\title{This function allows you to split a string and keep the delimeter}
\usage{
om_strsplit(x, split, type = "remove", perl = FALSE, ...)
}
\description{
This function allows you to split a string and keep the delimeter
}
