% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{om_textscore}
\alias{om_textscore}
\title{Use various text analysis tools}
\usage{
om_textscore(.data, method, text, verbose = T)
}
\arguments{
\item{.data}{dataset including text and a personal identifier (must be named `OMID`)}

\item{method}{which method(s) should be applied? Currently supports: "nrc", "bing", "afinn", "textblob", "vader", "sentiment_stanford", "sentimentr_polarity", "affect3d", "receptiveness", and "toxicity"}

\item{text}{specify the text variable which should be scored}
}
\description{
This function performs sentiment analysis and other text tools on any custom strings that you supply.
}
