% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{parse_feedback_at}
\alias{parse_feedback_at}
\title{This function parses feedback from AirTable (dat.par)}
\usage{
parse_feedback_at(raw_input)
}
\arguments{
\item{raw_input}{AirTable data: column FeedbackAnswers}
}
\description{
This function parses feedback from AirTable (dat.par)
}
