% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifehacks.R
\name{parse_lifehacks}
\alias{parse_lifehacks}
\title{Parse Lifehacks}
\usage{
parse_lifehacks(x, var)
}
\arguments{
\item{x}{LifeHack String}

\item{var}{How to name the variable}
}
\description{
Parses Lifehacks from GuidedTrack AND AirTable data
}
