% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{perc_improved}
\alias{perc_improved}
\title{Helper function to calculate percent of people who improved}
\usage{
perc_improved(all_pre, all_post, total, variable_code, reverse = F)
}
\arguments{
\item{all_pre}{Pre Data}

\item{all_post}{Post data}

\item{total}{N}

\item{variable_code}{variable code}
}
\description{
This function calculates percent of people who improved from pre to post/followup
}
