% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_helpers.R
\name{pipe_print}
\alias{pipe_print}
\title{This function allows you to print a message in a pipe chain}
\usage{
pipe_print(data, message)
}
\arguments{
\item{data}{Dataset pipethrough}

\item{message}{What message should be printed?}
}
\description{
This function allows you to print a message in a pipe chain
}
