% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{polar_measures}
\alias{polar_measures}
\title{Creates Political Polarization variables}
\usage{
polar_measures(final_dat, ProgTemp, ConTemp)
}
\arguments{
\item{final_dat}{Assessment data from AirTable}

\item{Q1}{Q1 variable}

\item{Q2}{Q2 variable}
}
\description{
This is lower-level function that belongs to om_construct measure. This function is not meant to be used outside of om_construct_measure.
Creates the following measures of Political Polarization
\itemize{
  \item Q14: Affective Polarization
  \item Q15: Ingroup
  \item Q16: Outgroup
  \item Q17: Ingroup vs. Outgroup Affective Polarization
}
}
