% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{remove_dups}
\alias{remove_dups}
\title{Remove duplicates from AirTable}
\usage{
remove_dups(cleaned_dat)
}
\arguments{
\item{cleaned_dat}{Duplicated data from AirTable}
}
\description{
Remove duplicates from AirTable
}
\section{Workflow}{

\code{\link{om_filter_data}} -> \code{\link{om_rescale}} -> \code{\link{om_clean_ppol}} -> \code{\link{om_construct_measures}} -> \code{\link{remove_dups}} -> \code{\link{om_gather}}
}

