% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtheme.R
\name{scale_color_om}
\alias{scale_color_om}
\title{A ggplot2 color palette}
\usage{
scale_color_om(...)
}
\description{
Blue : #2a98db
}
\details{
Purple : #3d4fa1

 Yellow : #e8df15

 Turquoise : #65c6c3

 Pink : #ec145b

 Green : #94cfa1

 Black : #414042

This object contains a palette with the official OpenMind colors
}
