% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_helpers.R
\name{strsplit}
\alias{strsplit}
\title{This function allows you to split a string and keep the delimeter}
\usage{
strsplit(x, split, type = "remove", perl = FALSE, ...)
}
\description{
This function allows you to split a string and keep the delimeter
}
