% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtheme.R
\name{theme_om}
\alias{theme_om}
\title{A ggplot2 theme}
\usage{
theme_om(legend_position = c(0.55, 0.93))
}
\arguments{
\item{legend_position}{Specify Legend position c(x, y)}
}
\description{
This function can be added to a ggplot2 call and creates the official OpenMind ggplot theme
}
