% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{withinSE}
\alias{withinSE}
\title{Get Within Standard Errors}
\usage{
withinSE(x, variable, WaveType)
}
\arguments{
\item{x}{data}

\item{variable}{specify which variable you want to calculate}

\item{WaveType}{which wave you want to compare (should be "Post" or "FollowUp")}
}
\description{
Get Within Standard Errors
}
