% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{merge_assessments}
\alias{merge_assessments}
\title{Merge Assessments}
\usage{
merge_assessments(v4, v5, v6)
}
\arguments{
\item{v4}{v4 data}

\item{v5}{v5 data}

\item{v6}{v6 data}
}
\description{
Merge Assessments
}
