% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{om_clean_par}
\alias{om_clean_par}
\title{Performs cleaning on ParticipantProgress data from AirTable}
\usage{
om_clean_par(dat.par, parse_feedback = F, ...)
}
\arguments{
\item{dat.par}{ParticipantProgress data from AirTable}

\item{parse_feedback}{Parse Feedback answers (Q1 to Q5 for Step 1 to 5). Default is \code{FALSE}.}

\item{...}{Arguments for select to get more from ParticipantProgress}
}
\description{
Creates the following measures
\itemize{
  \item StepTimes1 to StepTimes5: Duration in minutes to complete a step
  \item StepCorrect1 to StepCorrect5: Percentage of correct answers for each step
  \item FeedbackAnswers Q1 to Q5
}
}
