% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{om_clean_ppol}
\alias{om_clean_ppol}
\title{Creates several measures of Political Orientation}
\usage{
om_clean_ppol(app.dat)
}
\arguments{
\item{app.dat}{Assessment data from AirTable}
}
\description{
Creates the following measures of Political Orientation
\itemize{
  \item ppol_raw: a variable that merges Assessment V4 and V5.1 spelling of Political Orientation (D4)
  \item ppol: a factor variable ordered from "Very Progressive/left" to "Very Conservative/right". Excludes all other categories as NA (classical liberal etc.)
  \item ppol_num: numeric variable ranging from 1 "Very Progressive/left" to 7 "Very Conservative/right"
  \item ppol_cat: a factor variable which has two categories "Progressive" and "Conservative". The rest is NA.
}
}
\section{Workflow}{

\code{\link{om_filter_data}} -> \code{\link{om_rescale}} -> \code{\link{om_clean_ppol}} -> \code{\link{om_construct_measures}} -> \code{\link{remove_dups}} -> \code{\link{om_gather}}
}

