% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{om_construct_measures}
\alias{om_construct_measures}
\title{Constructs measures}
\usage{
om_construct_measures(x)
}
\arguments{
\item{x}{Assessment data from AirTable}
}
\description{
This is a higher-level function that uses both "polar_measures" and "calc_ih" to constuct various measures.
Creates the following variables:
\itemize{
  \item Q14: Affective Polarization
  \item Q15: Ingroup
  \item Q16: Outgroup
  \item Q17: Ingroup vs. Outgroup Affective Polarization
  \item Q18: Intellectual Humility
}
Function automatically accounts for Assessment Version 4 and 5/5.1.
}
