% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{om_download_at}
\alias{om_download_at}
\title{Download AirTable data}
\usage{
om_download_at(
  key,
  tables = c("AccessCodes", "ParticipantProgress", "InstructorSurveyV2",
    "TechnicalInquiries"),
  clean = F,
  file = NULL,
  v6.1 = F
)
}
\arguments{
\item{key}{key for AirTable API}

\item{tables}{specify which tables you want to download}

\item{clean}{clean dataset and construct measures (only works for Assessment V6 and V7)}

\item{file}{specify path to download to (only works for Assessment V6 and V7)}
}
\value{
a list with (several) dataframe(s)
}
\description{
This function downloads current AirTable data
}
\section{Currently allowed input}{

\itemize{
\item \code{"AssessmentV6"}
\item \code{"AssessmentV7"}
\item \code{"AccessCodes"}
\item \code{"ParticipantProgress"}
\item \code{"ParticipantProgress2"}
\item \code{"InstructorSurveyV2"}
\item \code{"TechnicalInquiries"}
}
}

\examples{
#Here is a code example that will download Assessment V6,
#clean it, save it into a folder called "Data" under
#Research and filter down to only include V6.1 data.
key <- readr::read_lines("../../Research/Projects/Keys/airtabler.txt")

assessmentv6 <- om_download_at(key,
                              tables = "AssessmentV6",
                              clean = TRUE,
                              file = "../../../Data/assessmentv6.1.csv",
                              v6.1 = TRUE)


#Here is another example code for downloading a clean version of Assessment v7:

assessmentv7 <- om_download_at(key,
                              tables = "AssessmentV7",
                              clean = TRUE)
}
