% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{om_filter_data}
\alias{om_filter_data}
\title{This function filters down a dataframe by the desired characteristics.}
\usage{
om_filter_data(app.dat, n_assessments = NULL, version = NULL,
  accesscode = NULL)
}
\arguments{
\item{app.dat}{Assessment data from AirTable}

\item{n_assessments}{{AssessmentsDone} Filter down to what asssesments? (Should be 1, 2 and/or 3)}

\item{version}{{AssessmentVersion} Filter down to what asssesment version?}

\item{accesscode}{{AccessCode} Filter down to specific AccessCodes}
}
\description{
This function filters down a dataframe by the desired characteristics.
}
