% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{om_gather}
\alias{om_gather}
\title{Gather OpenMind data}
\usage{
om_gather(.data, which_strings)
}
\arguments{
\item{.data}{Assessment data}

\item{which_strings}{a string indicating which variables should be parsed out}
}
\description{
This function will turn Assessment data into long format and
creates the following variables:
\itemize{
  \item Question: Q1Pre, Q2Pre, Q3Pre etc.
  \item Type: Pre, Post, or FollowUp
  \item Response: Values of the Question
  \item variable_code: Q1, Q2, Q3 etc.
}
}
