% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{om_summarize_comparisons}
\alias{om_summarize_comparisons}
\title{Conduct t-tests and calculate Cohen's d}
\usage{
om_summarize_comparisons(
  gathered_dat,
  aversion = "All",
  compare = c("PrePost", "PreFollow", "PrePostFollow")
)
}
\arguments{
\item{gathered_dat}{Assessment data as long format}

\item{aversion}{AssessmentVersion should be one of \code{"V4"}, \code{"V5/V5.1"} or \code{"All"}}

\item{compare}{With the `compare` argument you can specify either \code{"PrePost"}, \code{"PreFollow"} or both \code{c("PrePost", "PreFollow")} comparisons (the latter is the default).}
}
\description{
This is a higher-level function that uses "om_compare", "bind_questions" and "summarize_comparison" to calculate t-tests and Cohen's d on Assessment data.
}
