% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{om_summarize_comparisons}
\alias{om_summarize_comparisons}
\title{Conduct t-tests and calculate Cohen's d}
\usage{
om_summarize_comparisons(gathered_dat)
}
\arguments{
\item{gathered_dat}{Assessment data as long format}
}
\description{
This is a higher-level function that uses both "bind_questions" and "summarize_comparison" to calculate t-tests and Cohen's d on Assessment data.
}
