% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{om_ttest}
\alias{om_ttest}
\title{Run T-Tests on Long Format data}
\usage{
om_ttest(gathered_dat, comparison)
}
\arguments{
\item{gathered_dat}{Long format data}

\item{comparison}{Three possible comparisons "PrePost", "PreFollowUpT1T2" or "PreFollowUpT1T3"}
}
\description{
This function performs t-tests on v6 and v7 data
}
