% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{remove_dups}
\alias{remove_dups}
\title{Remove duplicates from AirTable}
\usage{
remove_dups(cleaned_dat)
}
\arguments{
\item{cleaned_dat}{Duplicated data from AirTable}
}
\description{
Remove duplicates from AirTable
}
